put tests/ testmachine in its own file
It's a short file yes.
but RunFileTest() in file.go is 305 lines long with many things inside,
defer, callbacks and what not. The simpler that file is, the easier it
will be to understand, maintain or refactor.

Since testmachine is very thematic, and short, there's little git blame
info to loose for testmachine.go.

We can now retain the git blame information on file.go while having it
seriously reduced in lines.
